fix: file and console sink not using the correct marshaller
