--- models/urn.go
@@ -1,10 +1,0 @@
-package models
-
-import "fmt"
-
-func NewURN(service, scope, kind, id string) string {
-	return fmt.Sprintf(
-		"urn:%s:%s:%s:%s",
-		service, scope, kind, id,
-	)
-}
